Changed player grounding logic
Grounding the player now uses 3 raycasts. The middle of which casts from the origin position of the player and the other two are cast from either side of the player's box collider.
This means that the player will be grounded pretty much as long as the bottom of the box collider is touching some part of the terrain on the specified LayerMask.