Cleaned up the code in PlayerMovementDataSO
Instead of having public auto properties to access private fields, I found a way to remove the private fields and just have the auto properties with private setters.

The problem before was that you cant serialize auto properties to the inspector by default. Instead of "[SerializeField]" if you put "[field: SerializeField]" it will now serialize the auto properties in the inspector. Same applies to other attributes like "Space" and "Header".