Changed "SampleScene" to act as project entrypoint
Changed the "SampleScene" to a project entry point. When the game loads it will load this scene first and we can modify this scene to suit our needs accordingly later.

For now it has a single game object with a script that will load a different scene (my test scene) on Awake()