Update GameEntryPoint scene and added a Bootstrapper script / prefab to additively load a new "Game Dependencies" scene.
- also added empty "Main Menu" scene to the project
- The GameEntryPoint scene will be the first thing that unity loads for the game, right now it contains the bootstrapper and another prefab to load a specific scene additively
-The bootstrapper will additively load a special "Game Dependencies" scene along side any currently loaded scenes that contains necessary content that must be present at all times for the game to function. Right now this includes the "Player Input" prefab and the "UI - Player HUD" prefab

You can add the bootstrapper prefab to any scene and it will provide Player Input and the PlayerHUD UI to that scene. It does not account for whether or not those prefabs already exist in some other scene so its best to use either the bootstrapper prefab or one of the other game dependent prefabs you need and not both at the same time.